% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model0.R
\name{generate_model0}
\alias{generate_model0}
\title{Generate a simple HIV epidemic model for use by phydynR}
\usage{
generate_model0(art_start = 2005)
}
\description{
* src : global reservoir
* i : infected and infectious
* z : infected and on ART
Transmission occurs at a rate beta(t) i(t) where beta is a linear
interpolating spline. The treatment rate i->z is initially zero but
increases linearly to 2015
}
