% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_sequences}
\alias{get_sequences}
\title{Title Get fasta sequences from PANGEA dataframe}
\usage{
get_sequences(all_data, country_name, outfile_name, add_unknown)
}
\arguments{
\item{all_data}{Dataframe containing all data}

\item{country_name}{String for country name of interest}

\item{outfile_name}{String with the output filenama to save sequences}

\item{add_unknown}{value "yes" or "no". If "yes" include all genders (females,
males and unknown), if "no" do not add sample in which gender is unknown.}
}
\value{
Fasta file for selected DNA sequences.
}
\description{
Function to select sequences from specific country and save these DNA sequences
as FASTA.
}
\examples{
#To Do.
}
