% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treestructure0.R
\name{getsubtres}
\alias{getsubtres}
\title{partition the tree into a set of clades within the desired size range}
\usage{
getsubtres(tr, minsize = 200, maxsize = 500, level = 0.25, NCPU = 1)
}
\description{
partition the tree into a set of clades within the desired size range
}
