% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treestructure0.R
\name{getsubtresRecursively}
\alias{getsubtresRecursively}
\title{partition the tree into a set of clades within the desired size range}
\usage{
getsubtresRecursively(
  tr,
  minsize = 200,
  maxsize = 500,
  level = 0.05,
  NCPU = 1,
  incrementLevel = 0.15
)
}
\description{
If initial partitioning step produces clades that are too big, will 
continue partitioning at hiher sign levels until 
desired size is reached
}
