% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model0summary.R
\name{m0R.t}
\alias{m0R.t}
\title{Compute R(t) by integrating over future hazard of removal and transmissions}
\usage{
m0R.t(s)
}
\description{
Compute R(t) by integrating over future hazard of removal and transmissions
}
