% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcFit0.R
\name{mcmcfit0.0}
\alias{mcmcfit0.0}
\title{Runs parallel mcmc chains in order to fit model0}
\usage{
mcmcfit0.0(
  BDTR,
  mod0,
  nc = 1,
  iterations = 10,
  likelihood = "PL2",
  continueFits = NULL
)
}
\arguments{
\item{BDTR}{A phydynR::DatedTree object}

\item{nc}{Integer number of CPUs and number of MCMC chains}

\item{iterations}{Length of MCMC chains}

\item{likelihood}{Switch for different coalescent likelihood approximations, passed to phydynR::colik}

\item{mode0}{A list with model objects generated by generate_model}
}
\description{
Priors are defined in the function
}
