% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{root_and_drop_tips}
\alias{root_and_drop_tips}
\title{Title Root phylogenetic tree and drop outgroup sequences}
\usage{
root_and_drop_tips(tree, outgroups)
}
\arguments{
\item{tree}{Phylogenetic tree of class phylo.}

\item{outgroups}{vector of string with names for outgroup sequences}
}
\value{
Rooted phylogenetic tree without outgroup sequences
}
\description{
Function to root a phylogenetic tree and drop outgroup sequences from the
tree.
}
\examples{
# TO DO
}
