% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_redundant_cgr.R
\name{rrcgr}
\alias{rrcgr}
\title{Prunes redundant close global reference sequences for subsequent
coalescent analysis.}
\usage{
rrcgr(dtr)
}
\arguments{
\item{dtr}{A rooted phylogenetic tree of class ape::phylo or treedater. CGR lineages and *only* CGR sequencs must include the 'CGR' string in the tip label.}
}
\description{
A CGR sequence is removed if it's closest 
neighbor in the tree is also a CGR. Thus the closest neighborr
for each CGR must be a non-CGR sequence
}
\examples{
load( system.file( 'data/env_treedater.rda', package = 'pangea') )
rrcgr( dtr_env2 ) -> tr 

}
