% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treedater0.R
\name{sa_treedater}
\alias{sa_treedater}
\title{Does a treedater analysis for gag, env or pol and usign the south africa data}
\usage{
sa_treedater(
  TREEFN = "sa_ml_gag.nwk",
  SRCMDFN = "SA_src_gag_metadata_mod.csv",
  PANGEAFN = "../../PANGEA_Extract_Volz_2019-04-16.csv",
  OG = c("K03454_gag", "AY371157_gag"),
  seqlen = 1476,
  NCPU = 8,
  outstem = "gag"
)
}
\description{
Does a treedater analysis for gag, env or pol and usign the south africa data
}
