% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setup.model.equations}
\alias{setup.model.equations}
\title{Setup Model Equations}
\usage{
setup.model.equations(demes, nondemes = NULL, rcpp = FALSE)
}
\arguments{
\item{demes}{a character vector naming the demes of the mathematical model.}

\item{nondemes}{a character vector naming the non demes of the mathematical model.}

\item{rcpp}{if TRUE, the expressions are interpreted as C code using the Rcpp package.}
}
\value{
this function returns a list containing the empty components
 (represented by zeros) to build the mathematical model.
 These components are the birth, death, migrations, total number of demes and
 non-demes of the model.
 \itemize{
     \item Birth is a vector or matrix describing the model birth rates;
     \item Death is a vector describing the model death rates;
     \item Migration is a vector or matrix describing the model migration
               rates.
 }
}
\description{
Function to setup the components of the mathematical model. See \code{\link[phydynR]{build.demographic.process}}
for more details
}
\examples{
demes <- c("gpm", "gpf", "msm", "src")
eqns <- setup.model.equations(demes)
}
\seealso{
\code{\link[phydynR]{build.demographic.process}}
}
