% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model0summary.R
\name{summary_model0}
\alias{summary_model0}
\title{Generate tables and plots for fits of model0}
\usage{
summary_model0(
  infns,
  art_start,
  burninpc = 0.5,
  ncpu = 1,
  outdir = NULL,
  numSamples = 100
)
}
\arguments{
\item{infns}{vector of filenames with MCMC fit in RDS format}

\item{burninpc}{proportion of mcmc to discard}

\item{ncpu}{}

\item{outdir}{where to store results; defaults to summary0}

\item{numSamples}{number of tractories to simulate}
}
\description{
Generate tables and plots for fits of model0
}
