% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model0summary.R
\name{trajectories_model0}
\alias{trajectories_model0}
\title{Make tfgys for multiple sets of input files}
\usage{
trajectories_model0(
  infns,
  art_start,
  mod0 = NULL,
  burninpc = 0.5,
  ncpu = 1,
  numSamples = 100,
  onlyBestChain = FALSE
)
}
\description{
Make tfgys for multiple sets of input files
}
